Added separated controller for API methods, configure serializer for display autocomplete hints
